Move clear_cache tests to test_utils.py
