+++ tests/test_prompt.py
@@ -1,13 +1,11 @@
 import os
 from typing import Optional
-import shutil
 
 import pytest
 from datasets import Dataset
 from pydantic import BaseModel
 
 from bespokelabs.curator import Prompter
-from bespokelabs.curator import clear_cache
 
 
 class MockResponseFormat(BaseModel):
@@ -139,54 +137,3 @@ def test_single_completion_no_batch(prompter: Prompter):
     assert isinstance(result, MockResponseFormat)
     assert hasattr(result, "message")
     assert hasattr(result, "confidence")
-
-
-@pytest.mark.test
-def test_clear_cache(tmp_path):
-    """Test that clear_cache removes all cached data."""
-    # Set up temporary cache directory
-    cache_dir = str(tmp_path)
-    os.environ["CURATOR_CACHE_DIR"] = cache_dir
-
-    # Create mock cache files and directories
-    metadata_db_path = os.path.join(cache_dir, "metadata.db")
-    os.makedirs(cache_dir, exist_ok=True)
-
-    # Create a mock metadata.db file
-    with open(metadata_db_path, 'wb') as f:
-        f.write(b'mock db content')
-
-    # Create mock cache directories and files
-    mock_cache_dir = os.path.join(cache_dir, "mock_cache")
-    os.makedirs(mock_cache_dir)
-    with open(os.path.join(mock_cache_dir, "cache.arrow"), 'w') as f:
-        f.write("mock cache data")
-
-    # Create a read-only file to test permission errors
-    readonly_file = os.path.join(cache_dir, "readonly.txt")
-    with open(readonly_file, 'w') as f:
-        f.write("readonly content")
-    os.chmod(readonly_file, 0o444)
-
-    # Verify cache exists
-    assert os.path.exists(metadata_db_path)
-    assert os.path.exists(mock_cache_dir)
-    assert os.path.exists(readonly_file)
-
-    # Clear cache
-    clear_cache(working_dir=cache_dir)
-
-    # Verify cache is cleared (even with readonly file)
-    assert not os.path.exists(metadata_db_path)
-    assert not os.path.exists(mock_cache_dir)
-    assert not os.path.exists(readonly_file)
-
-    # Test clearing non-existent directory (should not raise error)
-    non_existent_dir = os.path.join(tmp_path, "non_existent")
-    clear_cache(working_dir=non_existent_dir)
-
-    # Test using environment variable
-    os.environ["CURATOR_CACHE_DIR"] = str(tmp_path)
-    os.makedirs(os.path.join(str(tmp_path), "env_var_test"))
-    clear_cache()  # Should use CURATOR_CACHE_DIR
-    assert len(os.listdir(str(tmp_path))) == 0
